requirements.txt
asgiref==3.8.1
decouple==0.0.7
Django==4.2.22
django-rest-framework==0.1.0
djangorestframework==3.16.0
sqlparse==0.5.3
typing_extensions==4.14.0
